Simplify run_impl for potential_temperature diagnostic
This commit simplifies the interface for the potential temperature
diagnostic following reviewer comments.  It removes the run_diagnostic
structure from the potential_temperature header file and simply runs the
Kokkos parallel loop directly in run_impl.

This also requires us to expose run_impl as a public member of the class.

There is some cleanup to the potential temperature test, creating a
separate header file `diagnostic_test_util.hpp` which holds the boiler
plate test utilities that can be used by all subsequent tests.

Finally, this commit removes two old files that should have never been
committed and are not used - diagnostics_interface.*
