+++ components/scream/src/diagnostics/potential_temperature.hpp
@@ -43,49 +43,17 @@ public:
   // Set the grid
   void set_grids (const std::shared_ptr<const GridsManager> grids_manager);
 
-  // Actual diagnostic calculation 
-  struct run_diagnostic_impl {
-    run_diagnostic_impl() = default;
-    // Functor for Kokkos loop
-    KOKKOS_INLINE_FUNCTION
-    void operator() (const int& idx) const {
-      const int icol  = idx / m_npack;
-      const int jpack = idx % m_npack;
-      const Spack& T_mid_ij(T_mid(icol,jpack));
-      const Spack& p_mid_ij(p_mid(icol,jpack));
-      output(icol,jpack) = PF::calculate_theta_from_T(T_mid_ij,p_mid_ij);
-    }
-    int m_ncol, m_npack;
-    view_2d_const T_mid;
-    view_2d_const p_mid;
-    view_2d       output;
-    // assign variables to this structure
-    void set_variables(const int ncol, const int npack,
-      const view_2d_const& pmid_, const view_2d_const& tmid_, const view_2d& output_)
-    {
-      m_ncol = ncol;
-      m_npack = npack;
-      // IN
-      T_mid = tmid_;
-      p_mid = pmid_;
-      // OUT
-      output = output_;
-    }
-  }; // struct run_diagnostic_impl
-
+  // run_impl needs to be public so we can use a Kokkos LAMBDA in it.
+  void run_impl        (const int dt);
 protected:
 
   // The three main overrides for the subcomponent
   void initialize_impl (const RunType run_type);
-  void run_impl        (const int dt);
   void finalize_impl   ();
 
   // Keep track of field dimensions
   Int m_num_cols; 
   Int m_num_levs;
-
-  // Structure to run the diagnostic
-  run_diagnostic_impl  run_diagnostic;
 
 }; // class PotentialTemperatureDiagnostic
